1、提交本地环境

+++ app/Http/Controllers/admin/CatesController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\CatesRequest;
 use App\Models\Cates;
 use DB;
+use App\Models\Goods;
 
 class CatesController extends Controller
 {
@@ -41,11 +42,11 @@ class CatesController extends Controller
         if($request->has('search')){
             $search = $request->input('search');
 
-            $cates = Cates::where('classname','like','%'.$search.'%')->paginate(2)->appends($request->input());
-            // dump($cates);
-            return view('admin.cates.index',['title'=>'分类列表','cates'=>$cates]);
+            $cates = Cates::where('classname','like','%'.$search.'%')->paginate(5)->appends($request->input());
+            $count = count($cates);
+            return view('admin.cates.index',['title'=>'分类列表','cates'=>$cates,'count'=>$count]);
         }else{
-            $cates = Cates::select('id','pid','classname','path','status',DB::raw("concat(path,',',id) as paths"))->orderBy('paths','asc')->paginate(3);
+            $cates = Cates::select('id','pid','classname','path','status',DB::raw("concat(path,',',id) as paths"))->orderBy('paths','asc')->paginate(5);
             foreach($cates as $k=>$v){
                 //统计逗号出现的次数
                 $i = substr_count($v->path,',');
@@ -208,11 +209,17 @@ class CatesController extends Controller
      */
     public function delete($id)
     {
-        $res = Cates::onlyTrashed()->where('id',$id)->forceDelete();
+        $goods = Goods::where('gcid',$id)->get();
+        if(empty($goods[0])){
+            $res = Cates::onlyTrashed()->where('id',$id)->forceDelete();
           if($res){
             return redirect('/admin/cates')->with('success','永久删除成功！！');
+            }else{
+                return back()->with('error','永久删除失败！！');
+            }
         }else{
-            return back()->with('error','永久删除失败！！');
+            return back()->with('error','分类下存在商品,请勿删除！！');
         }
+        
     }
 }
